Add new options class and provide a callback to retrieve it, rather than requiring config to be provided immediately.
